Prototype/Prototype 1/arc_ai_net_sim/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.30.0
pydantic==2.6.4
httpx==0.27.0
cryptography==42.0.5
PyNaCl==1.5.0
python-dotenv==1.0.1
aiosqlite==0.19.0
sqlalchemy==2.0.29
alembic==1.13.1
pytest==8.1.1
pytest-asyncio==0.23.5
pytest-cov==4.1.0
typer==0.9.0
orjson==3.10.6
jinja2==3.1.4
python-multipart==0.0.9

